adding comments to CompressionUtils.h

+++ compression/src/CompressionUtils.h
@@ -18,6 +18,10 @@ inline T getThresholdForTopK(const std::vector<T>& values, uint32_t sketch_size,
                              seed_for_sampling);
 }
 
+/*
+ * Rather than getting an exact threshold, we sample a number of points from the
+ * values array and gets an estimate for topk.
+ */
 template <class T>
 inline T getThresholdForTopK(const T* values, uint32_t size,
                              uint32_t sketch_size,
@@ -27,35 +31,43 @@ inline T getThresholdForTopK(const T* values, uint32_t size,
 
   uint32_t min_samples = 20;
 
-  // there will be scenarios when num_samples takes very small values such as 1,
-  // hence, we ensure that we sample at least 20 points
+  /*
+   * There will be scenarios when num_samples takes very small values such as 1,
+   * hence, we ensure that we sample at least 20 points.
+   */
+
   if (num_samples < min_samples) {
     num_samples = static_cast<uint32_t>(std::min(min_samples, size));
   }
 
-  // sketch_size and size are both uint32_t and hence, we multiply by 1.0 to
-  // upcast the division
+  /*
+   * sketch_size and size are both uint32_t and hence, we multiply by 1.0 to
+   * upcast the division.
+   */
+
   uint32_t topK = static_cast<uint32_t>(1.0 * num_samples * sketch_size / size);
 
   std::vector<T> sampled_gradients(num_samples, 0);
 
-  // changing rand is very important
   std::mt19937 gen(seed_for_sampling);
   std::uniform_int_distribution<> distrib(0, size - 1);
   for (uint32_t i = 0; i < num_samples; i++) {
     sampled_gradients[i] = std::abs(values[distrib(gen)]);
   }
 
-  // threshold is an estimate for the kth largest element in the gradients
-  // matrix
-
   std::nth_element(sampled_gradients.begin(),
                    sampled_gradients.begin() + num_samples - topK,
                    sampled_gradients.end());
+
+  // threshold is an estimate for the kth largest element in the gradients
+  // matrix
   T threshold = sampled_gradients[num_samples - topK];
   return threshold;
 }
 
+/*
+ * Given the number of chunks, we split a vector into n almost equal chunks
+ */
 template <class T>
 std::vector<std::vector<T>> splitVector(const std::vector<T>& vec, size_t n) {
   std::vector<std::vector<T>> out_vec;
@@ -66,7 +78,7 @@ std::vector<std::vector<T>> splitVector(const std::vector<T>& vec, size_t n) {
   size_t begin = 0;
   size_t end = 0;
 
-  for (size_t i = 0; i < std::min(size_t(n), vec.size()); ++i) {
+  for (size_t i = 0; i < std::min(n, vec.size()); ++i) {
     end += (remain > 0) ? (length + !!(remain--)) : length;
 
     out_vec.emplace_back(
